Move main JS
